data/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi>=0.78.0,<1
pydantic>=1.8.0,<2.0.0
uvicorn>=0.17.6,<1
requests==2.28.1
httpx==0.23.3
asyncio==3.4.3
aiohttp==3.8.4
elasticsearch==8.6.2
fastapi_websocket_pubsub==0.3.1 
pyee==9.0.4
python-dateutil<3.0.0,>=2.8.2
# ------------------------- #
# DEV - TODO: use env
# ------------------------- #

# ------------------------- #
# END DEV - TODO: use env
# ------------------------- #